Added the ability to specify the width of each file when generating interleaved copies
You specify the number of bytes to store to each file before switching
to the next one. If you had 16-bit wide flash chips then you would set
it to two if the bus was 32-bits wide. This is done in a strange fashion
to make sure that it does what it says. Nevermind the fact that it is
less efficient this way but correct.